+++ index.html
@@ -48,6 +48,7 @@
     </div>
   </body>
 
-  <script src="scriptHW9.js">
+
+<script type="module" src="scriptHW9.js">
   </script>
 </html>
